Add lambda at edge permission to create log groups
This is stupid, because I want to set the expiry date for groups. But lambda at edge
creates a log stream in the region where the lambda was invoked. Because of this, I would
have to deploy a cfn template in all regions or create a custom resource. And I dont feel
like doing all that right know for some logs.